chore : 이미지 업로드 로직 삭제 및 수정
