+++ api/build.gradle
@@ -42,8 +42,6 @@ dependencies {
     annotationProcessor "com.querydsl:querydsl-apt:5.0.0:jakarta"
     annotationProcessor "jakarta.annotation:jakarta.annotation-api"
     annotationProcessor "jakarta.persistence:jakarta.persistence-api"
-    // AWS SDK
-    implementation 'com.amazonaws:aws-java-sdk-s3:1.12.749'
     // Test
     testImplementation 'org.springframework.boot:spring-boot-starter-test' // Spring Boot Test
     testImplementation 'org.springframework.security:spring-security-test' // Spring Security Test
