+++ api/src/main/java/org/meotppo/webti/dto/PropensityAnalysis/PropensityProfileResponseDto.java
@@ -1,4 +1,4 @@
-package org.meotppo.webti.dto.PropensityAnalysis;
+package org.meotppo.webti.dto.propensityAnalysis;
 
 import org.meotppo.webti.domain.entity.type.MbtiType;
 import org.meotppo.webti.dto.file.ImageDto;
@@ -11,13 +11,13 @@ public class PropensityProfileResponseDto {
     private String result;
     private String description;
     private MbtiType mbtiType;
-    private ImageDto image;
+    private ImageDto imageDto;
 
     @Builder
-    public PropensityProfileResponseDto(String result, String description, MbtiType mbtiType, ImageDto image) {
+    public PropensityProfileResponseDto(String result, String description, MbtiType mbtiType, ImageDto imageDto) {
         this.result = result;
         this.description = description;
         this.mbtiType = mbtiType;
-        this.image = image;
+        this.imageDto = imageDto;
     }
 }
