+++ api/src/main/java/org/meotppo/webti/service/file/FileService.java
@@ -1,46 +1,50 @@
 package org.meotppo.webti.service.file;
 
-import java.io.IOException;
-import java.util.List;
-
-import org.meotppo.webti.common.FileStore;
 import org.meotppo.webti.domain.entity.jpa.file.Image;
 import org.meotppo.webti.domain.repository.jpa.file.ImageRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import org.springframework.web.multipart.MultipartFile;
 
 import lombok.RequiredArgsConstructor;
+
+import java.util.UUID;
 
 @Service
 @RequiredArgsConstructor
 @Transactional(readOnly = true)
 public class FileService {
-    private final FileStore fileStore;
     private final ImageRepository imageRepository;
 
     @Transactional
-    public List<Image> storeFiles(List<MultipartFile> multipartFiles) throws IOException { // 여러 개 이미지 업로드 서비스 로직
-        List<Image> storedFiles = fileStore.storeFiles(multipartFiles);
-        imageRepository.saveAll(storedFiles);
-        return storedFiles;
-    }
-
-    @Transactional
-    public Image storeFile(MultipartFile multipartFile) throws IOException { // 단일 이미지 업로드 서비스 로직
-        Image storedFile = fileStore.storeFile(multipartFile);
-        if (storedFile != null) {
-            imageRepository.save(storedFile);
-        }
-        return storedFile;
-    }
-
-    @Transactional
     public Image storeImage(String url) { // 업로드 된 단일 이미지 저장 서비스 로직
-        Image storedFile = fileStore.storeImage(url);
+        if (url.isEmpty()) {
+            return null;
+        }
+
+        String originalurl = url;
+        String originalFilename = url.substring(url.lastIndexOf("/") + 1); // url에서 파일명 추출
+        String storeFileName = createStoreFileName(originalFilename); // 파일명 생성
+        
+        Image storedFile = Image.builder()
+                .originalFileName(originalFilename)
+                .storeFileName(storeFileName)
+                .url(originalurl)
+                .build();
+
         if (storedFile != null) {
             imageRepository.save(storedFile);
         }
         return storedFile;
+    }
+
+    private String createStoreFileName(String originalFilename) {
+        String ext = extractExt(originalFilename);
+        String uuid = UUID.randomUUID().toString();
+        return uuid + "." + ext;
+    }
+
+    private String extractExt(String originalFilename) {
+        int pos = originalFilename.lastIndexOf(".");
+        return originalFilename.substring(pos + 1);
     }
 }
